// src/util/types.ts
export type TopicSource = 'telegram' | 'discord' | 'donation' | 'log' | 'manual' | 'youtube' | 'unknown'
// ...
export interface EpisodeMashup {
    id: string
    type: 'mashup'
    topic?: Topic
    date: number
    characters: string[]
    videoId: YoutubeVideoID
    dialogData?: undefined
    mashupData: EpisodeMashupData
    isReady: boolean
    progress: number
    progressMax: number
    donations: Donation[]
}
export interface Topic {
    text: string
    id: string
    date: Date
    priority: boolean
    author?: string
    source: TopicSource
}
export interface YoutubeVideoID extends String {
    length: 11
}
export interface EpisodeMashupData {
    audio: {
        mix: string
        vocals: string
    }
    events: EpisodeMashupEvent[]
    bpm: number
    beatOffset: number
    lengthSeconds: number
    lengthBeats: number
}
export interface EpisodeMashupEvent {
    type: 'dropStart' | 'dropEnd' | 'setSinger'
    time: number // in beats
    content?: string
}
export interface EpisodeAction {
    type: 'line'|'narration'
    id: string
    character?: string
    emotion?: string
    content: string
}
export interface EpisodeVoiceLine {
    character?: string
    fetched: boolean
    url: string
    type: 'wav'|'mp3'|'ogg'
}
export interface Donation {
    id: string
    author: string
    message?: string
    amount: number
    currency: "RUB"|"USD"
    date: Date
}